refactor: trim unnecessary exports
