+++ packages/kit/src/index.ts
@@ -5,7 +5,6 @@ export { Fault } from '@stops/Fault';
 export { Silence } from '@stops/Silence';
 export { CustomId } from '@customId/CustomId';
 export { paginate } from '@pagination/paginate';
-export { pageCursor, type PageCursor } from '@pagination/cursor';
 export { type PageView } from '@pagination/PageView';
 
 /** Package version */
