+++ packages/kit/src/internal.index.ts
@@ -1,7 +1,7 @@
 export { setBotColor } from '@src/botColorHolder';
 
 export { prefixOf, decodeFor, type AnyCustomId } from '@customId/CustomId';
-export { PAGE_MAX } from '@pagination/cursor';
+export { PAGE_MAX, pageCursor, type PageCursor } from '@pagination/cursor';
 export { ComponentDefsKey, type HasComponentDefs } from '@customId/routing';
 export type { DecodedParams } from '@customId/Field';
