+++ packages/kit/src/pagination/cursor.ts
@@ -11,23 +11,13 @@ export const PAGE_MAX = 1_000_000;
 // @internal
 export const SLOT_MAX = 4;
 
-/**
- * A page cursor for a paginator. Use `PageCursor<'bans'>` or extract it from `typeof paginator.cursor`.
- *
- * @typeParam Prefix - The route prefix the dispatcher matches.
- */
+/** Page cursor for a paginator. */
 export type PageCursor<Prefix extends string> = CustomId<
     Prefix,
     { page: CustomIdField<number>; slot: CustomIdField<number> }
 >;
 
-/**
- * Build the page cursor for a paginator. The bounded `page` and `slot` fields encode under Discord's
- * 100-char customId cap.
- *
- * @param prefix - The route prefix the dispatcher matches.
- * @returns A {@link PageCursor} for `prefix`.
- */
+/** Build the page cursor for a paginator. */
 export function pageCursor<Prefix extends string>(prefix: Prefix): PageCursor<Prefix> {
     return new CustomId(prefix).int('page', 0, PAGE_MAX).int('slot', 0, SLOT_MAX);
 }
