+++ packages/seedcord/src/pagination/index.ts
@@ -1,7 +1,4 @@
 export { Paginator } from './Paginator';
 export { ArraySource, CursorSource } from './sources';
 
-export type { PageContext } from './PageContext';
-export type { PageSource } from './sources';
-export type { Renderable, ItemRender, PageRender } from './render';
-export type { PaginatorControls, ControlKey, ControlCosmetics } from './controls';
+export type { PaginatorControls } from './controls';
