+++ packages/seedcord/src/pagination/render.ts
@@ -4,7 +4,8 @@ import { TextDisplayBuilder } from 'discord.js';
 import { Controls } from './controls';
 
 import type { ControlKey, PaginatorControls } from './controls';
-import type { PageCursor, PageView } from '@seedcord/kit';
+import type { PageView } from '@seedcord/kit';
+import type { PageCursor } from '@seedcord/kit/internal';
 import type { ReplyResponse, V2Component } from '@seedcord/types';
 import type { ButtonBuilder, ActionRowBuilder } from 'discord.js';
